requirements.txt
fastapi
Jinja2
python-multipart
python-dotenv
requests
uvicorn
